requirements.txt
flask==3.0.0
pandas==2.1.1
python-dotenv==1.0.0
google-generativeai==0.3.2
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
pandasai==3.0.0b18
pandasai-google 